Fixed path to reportIdP.php, added basic auth
